Insertamos los Servlets para Validar usuario y Cerrar la sesión.

+++ web/pensionistas.jsp
@@ -32,7 +32,6 @@
                     f_inicio = contratoPensionista.getFechaInicio();
                     f_final = contratoPensionista.getFechaFin();
                 
-            
             }      
         %>
 <div class="container">
@@ -75,7 +74,6 @@
             <tbody>
                 <%
                     int count = 0;
-                    //Pensionistadao dao = new PensionistadoaImpl();
 
                     for (Listar_pensionista per : dao.ListarPensionista()) {
                         count++;
@@ -92,7 +90,7 @@
                     <td><%=per.getFechaInicio().substring(0,10)%></td>
                     <td><%=per.getFechaFin().substring(0,10)%></td>
                     <td><p><a class="btn btn-primary" title="Renovar Contrato del Pencionista" href="renovar.jsp?id=<%=per.getIdPersona()%>&nombres=<%=per.getNombres()%>&apellidos=<%=per.getApellidos()%>&dni=<%=per.getDni()%>&ncelular=<%=per.getNCelular()%>&direcciones=<%=per.getDireccion()%>" role="button"><i class="glyphicon glyphicon-refresh"></i></a></p></td>
-                    <td><p><a class="btn btn-danger" role="button" href="pensionistas.jsp?opcion=delete&id=<%=per.getIdContrato()%>"><i class="glyphicon glyphicon-trash"></i></a></p></td>
+                    <td><p><a class="btn btn-danger" title="Eliminar" onclick="if(!confirm('Esta seguro de eliminar a <%=per.getNombres()%><%=per.getApellidos()%>'))return false" role="button" href="pensionistas.jsp?opcion=delete&id=<%=per.getIdContrato()%>"><i class="glyphicon glyphicon-trash"></i></a></p></td>
                 </tr>
                 <%}%>
                 <%=mensaje%>
